Weighted least-squares routine added to mathutil
polyfitw was added to mathutil.  This function works similarly to polyfit in
the numpy/scipy packages, except that it permits the user to input weights to
the data points.  This is useful for including measurement error of the data
points.

Also added linear_leastsq to mathutil, which performs a simple linear
least-squares fit, but conveniently returns statistical errors associated with
the fit.

darcs-hash:20100316201041-99061-8b4cdc77a6e368d63833bfe0813316cfde44ca75.gz